rename: friends 페이지 폴더 정리
