+++ src/components/friends/AddFriend/CopyMyIdButton.tsx
