+++ src/components/friends/AddFriend/FriendRequest.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import { useRecoilValue } from 'recoil'
 import { FlexBox } from '@components/layouts/FlexBox'
 import { Text } from '@components/atoms/Text'
-import FriendsListDItem from './FriendsListDItem'
+import FriendsListDItem from '../FriendsList/FriendsListDItem'
 import { authState } from '@libs/store/auth'
 import { FriendsApi } from '@utils/apis/friends/FriendsApi'
 import { Spacing } from '@components/atoms/Spacing'
