+++ src/components/friends/AddFriend/SearchedFriendItem.tsx
