+++ src/components/friends/AddFriend/ShareMyId.tsx
