+++ src/components/friends/AddFriend/ShareMyIdButton.tsx
