+++ src/components/friends/FriendsList/FriendsListBItem.tsx
