+++ src/components/friends/FriendsList/FriendsListCItem.tsx
