+++ src/components/friends/FriendsList/NewFriends.tsx
