+++ src/components/friends/PresentRecommend/PresentRecommend.tsx
@@ -9,10 +9,10 @@ import { ItemFilter } from '@assets/icons/ItemFilter'
 import { Text } from '@components/atoms/Text'
 import { FriendsApi } from '@utils/apis/friends/FriendsApi'
 import { GiftFilter } from '@components/atoms/GiftFilter'
-import SortItem from './bottomsheet/SortItem'
+import SortItem from '../bottomsheet/SortItem'
 import { useRecoilState, useRecoilValue } from 'recoil'
 import { FilterState, isFilterTypeState, PriceState } from '@libs/store/present'
-import PriceFilter from './bottomsheet/PriceFilter'
+import PriceFilter from '../bottomsheet/PriceFilter'
 import { PriceFilterIcon } from '@assets/icons/PriceFilterIcon'
 import { theme } from '@styles/theme'
 import BottomSheet from '@components/atoms/BottomSheet'
