+++ src/components/friends/SearchFriend/FriendNotFound.tsx
