+++ src/pages/friends/AddFriend.tsx
@@ -3,9 +3,9 @@ import { useRecoilValue } from 'recoil'
 import { friendState } from '@libs/store/friend'
 import { useSetFriendRecoilState } from '@libs/hooks/useSetFriendRecoilState'
 import { Spacing } from '@components/atoms/Spacing'
-import FriendRequest from '@components/friends/FriendRequest'
-import ShareMyId from '@components/friends/ShareMyId'
-import AddFriendById from '@components/friends/AddFriendById'
+import FriendRequest from '@components/friends/AddFriend/FriendRequest'
+import ShareMyId from '@components/friends/AddFriend/ShareMyId'
+import AddFriendById from '@components/friends/AddFriend/AddFriendById'
 
 const AddFriend = () => {
   const friendStateData = useRecoilValue(friendState)
