+++ src/pages/friends/SearchFriends.tsx
@@ -4,8 +4,8 @@ import { useRecoilValue } from 'recoil'
 import { authState } from '@libs/store/auth'
 import { SearchInput } from '@components/atoms/Input/SearchInput'
 import { Spacing } from '@components/atoms/Spacing'
-import FriendsListBItem from '@components/friends/FriendsListBItem'
-import FriendNotFound from '@components/friends/FriendNotFound'
+import FriendsListBItem from '@components/friends/FriendsList/FriendsListBItem'
+import FriendNotFound from '@components/friends/SearchFriend/FriendNotFound'
 import { FriendsApi } from '@utils/apis/friends/FriendsApi'
 import { FriendsType } from '@utils/apis/friends/FriendsType'
 import styled from '@emotion/styled'
